+++ src/main/java/roomescape/dto/ReservationTimeResponse.java
@@ -2,7 +2,7 @@ package roomescape.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeResponse (
         long id,
