+++ src/test/java/roomescape/service/FakeReservationDao.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationDao;
-import roomescape.entity.Reservation;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationDao;
 
 public class FakeReservationDao implements ReservationDao {
