+++ src/test/java/roomescape/service/FakeReservationTimeDao.java
@@ -4,8 +4,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicLong;
-import roomescape.dao.ReservationTimeDao;
-import roomescape.entity.ReservationTime;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeDao;
 
 public class FakeReservationTimeDao implements ReservationTimeDao {
